Allow to mount Admin namespace on a specific path.
When Ektar is mounted at "/admin", navigating to users gives a path like "/admin/admin/users" which is odd.

With the new admin_path setting a path can be especified, ex: mounting Ektar at "/admin" and the "admin_path" setting set as "/" will render the url for users as "/admin/users".